Updated the callback system.
git-svn-id: https://svn.code.sf.net/p/tricklesync/code/trunk@31 dd4b9e15-e244-0410-b754-811dee3e6019

+++ SyncSimulationFramework/SimulationCallbackProtocol.h
@@ -10,5 +10,5 @@
 #include <Carbon/Carbon.h>
 
 @protocol SimulationCallbackProtocol
-- (void)setPercentCompleted:(int)value;
+- (void)setPercentCompleted:(double)value;
 @end
